requirements.txt
streamlit>=1.38.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-chroma>=0.1.0
langchain-google-genai>=2.0.0
pypdf
chromadb
python-dotenv
pydantic
google-generativeai
lark